FAC-363: minor moving
