+++ flickit-assessment-users/src/test/java/org/flickit/assessment/users/application/port/in/spaceuseraccess/InviteSpaceMemberUseCaseParamTest.java
@@ -1,8 +1,9 @@
-package org.flickit.assessment.users.application.port.in.spaceaccess;
+package org.flickit.assessment.users.application.port.in.spaceuseraccess;
 
 import java.util.UUID;
 
 import jakarta.validation.ConstraintViolationException;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.InviteSpaceMemberUseCase;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
